Support GPORCA building in autoconf
The --enable-orca autoconf switch was setting up some of the required
scaffolding for building GPDB with GPORCA but not all that is needed
to get a working GPDB with GPORCA out of the box. The makefiles were
assuming builds in gpAux/ext which isn't the case when building GPORCA
from source out of the OSS repository. The main changes in this patch:

* Adds a --with-xerces=DIR autoconf setting which is used to identify
  the location of the patched Xerces-C library needed by GPORCA. Once
  GPORCA no longer requires a patched version this can be removed. If
  using GPXerces-C in the base install path or a patched version of
  Xerces-C in the system folder then this switch isn't required. This
  is _only_ for building with a Xerces-C installation outside the main
  system folder.
* Checks for all used header files of the required libraries.
* Checks for all require libraries.
* Sets the linking of the GPORCA libraries properly in backend/gpopt.
* Installs the libdxltranslators library on make install and supports
  make uninstall as well.

All of this still requires GPORCA properly built and installed as per
the build instructions in the GPORCA repository. Should the libs from
GPORCA be installed with a --prefix then this needs to be added to
GPDB autoconf invocation with the --with-libraries=DIR option.