Use finegrained .out ignoring in the output directory .gitignores
We already had a mostly up to date list in the directory specific
.gitignore files but the toplevel .gitignore contained a *.out
which took precedence (lower level files override higher level
files so removing *.out will allow for the src/test/regress/* files
to override the now removed .out file blocking). The reason for
not ignoring *.out is that we have quite a few tracked .out files
in the repository which we don't want to ignore changes to.

Also update the files to match the current set of tests since there
were a few missing.