Fix "record type not registered" error, when a subquery returns a row var.
PostgreSQL relies on ExecEvalWholeRowVar() to assign a typmod for record
types, but that's too late in GPDB.  This was revealed by this regression
test that was added to upstream in version 8.3:

select q from (select max(f1) from int4_tbl group by f1 order by f1) q;

That test's purpose was to test something else, but it happened to trigger
this GPDB-specific bug. No test case included in this patch, because we'll
get that query later, as part of the 8.3 merge.