s3ext: fix memleaks and a forwarding null bug
This commit fixes issues below:

extractContent() might forward NULL to marker.
ListBucket() fogets to delete keylist when errors happen.
contents in ListBucketResult are not cleared.
downloader is not destroyed in unit tests.
get_opt_s3() and truncate_options() in unit tests need to be freed.
xmlNodeGetContent() needs to be freed.
xml.ctxt->myDoc needs to be freed.
function for the XML library needs to be cleared.

Thanks to Kuien and Valgrind.