Fix rescan support for external scan nodes.
Instead of pushing the responsibility of rescanning down to each different
kind of external table implement rescanning in fileam.c in a generic
fashion, by closing and reopening the underlying "url". This gets us rescan
support for custom and EXECUTE-type external tables, which was missing
before. The missing support caused a crash if you have a plan that requires
rescanning, as the one added in the regression test.