Remove outdated and unneeded method overrides in gpconfigurenewsegment.
In gpconfigurenewsegment, the class ConfExpSegCmd inherits from the class Command.
Command uses a variable of type ExecutionContext which was updated back in December
2011 to include passing the Command object itself to the execute(), interrupt(), and
cancel() ExecutionContext methods. The author of that patch did not update the
gpconfigurenewsegment.ConfExpSegCmd class which overrides those methods. It doesn't
make sense to even override them in the first place so they should just be removed.