Honour PLAIN storage also when constructing a memtuple.
A function might legitimately assume that an argument's varlen datum always
has a 4-byte header, if the datatype is marked as 'plain', and crash if we
then pass it a datum with 1-byte header, because it was packed in a
memtuple. I bumped into this while working on the PostgreSQL 8.3 merge,
because the merge brought us one such function: ts_rewrite().