Fix dispatching of operator family and class DDL commands.
When I merged the operator family patch, I missed dispatching the new DDL
commands to segments. Because of that, the segments didn't have information
about operator families. Some operator families would be greated implicitly
by CREATE OPERATOR CLASS, but you wouldn't necessarily get the same
configuration of families and classes as in the master. Things worked pretty
well despite that, because operator families and classes are used for
planning, and planning happens in the master. Nevertheless, we really should
have the operator family information in segments too, in case you run
queries in maintenance mode directly on the segments, or if you execute
functions in segments that need to execute expression that depend on them.
Also, there were no regression tests for the new DDL commands.