Remove redundant test case from qp_functions test file.
The test with stress_test() function (and accompanying tables) was created
and executed once. Then it was dropped, and recreated, and then executed
two times. Executing the same function twice might reveal bugs in plan
caching, so I kept that (although TBH we have better coverage for that
elsewhere). But I don't see the point of dropping and recreating it in
between: surely it's good enough to just create the function once, and
execute it twice.

This reduces the runtime of qp_functions test by about 1/3 (from 3 minutes
to 2 minutes on my laptop).