Remove dead code in outfuncs.c
The reason to remove them is because: print_variable_fields is initialized
to true, and only set to false in outfuncs_workfile_mgr_init() function.
But outfuncs_workfile_mgr_init() is never called. So we can remove all of:
print_variable_fields, outfuncs_workfile_mgr_init(), outfuncs_workfile_mgr_end(),
and all the if() tests depending on print_variable_fields.

Co-authored-by: Kuien Liu and Heikki Linnakangas