Extend mapred reduce function test with builtin reducer
The agebracket.yml mapred test case is testing the reduction with
a simple reducer summing the values. GPMapred has a builtin SUM
reducer however which will yield the same result, extend the test
with a copy of the agebracket.yml test using the builtin ensuring
that the output is identical.