Remove -O3 -funroll-loops -finline-limit=1800 from "standard" gcc options.
The "standard" gcc options in gpAux/Makefile, used to do a full build of
the installers, were using dubious compiler flags. -O3 enables some extra
optimizations that might be beneficial in some cases, but it also makes
binaries larger which can in fact hurt performance. Likewise, aggressive
inlining can improve performance in hot spots, but when applied
indiscriminately, it makes binaries larger for little gain. For lack of
better evidence, let's trust the compiler defaults. This makes the build
faster, and the binaries smaller. If we find out that these options are
beneficial for some specific functions or source files, let's add them back
to those special cases.

Also change the default from -O3 to -O2 in configure.in. That is the
default in the upstream, too.