Explicitly SET gp_select_invisible to false instead of using RESET
As outlined in https://github.com/greenplum-db/gpdb/issues/685 we
have a problem with resetting GUCs in the backends when the GUC
was previously set in the per connection parameter. Explicitly set
this GUC to false since that's the intention with the RESET commands
in these cases to increase robustness of the tests.