Avoid "HEAP_MOVED_OFF was expected" errors in VACUUM FULL.
If a page has hint bits set, but the buffer has not been marked as dirty,
and it gets evicted between the 1st and 2nd vacuum pass, the 2nd pass
gets upset. That can't happen in the upstream, as setting a hint bit
always marks the buffer as dirty, but that is not guaranteed in GPDB, because
of gp_disable_tuple_hints.