Move memoryAccount field from PlannedStmt to QueryDesc.
This isn't a problem right now, but with PostgreSQL 8.3, PlannedStmt is
supposed to be immutable, because plan caching works by saving the
PlannedStmt. So let's tidy this up, before we merge that from the upstream.
There are more GPDB-added fields in there that are modified, but this is a
start.