Split bitmap-index related tests a separate test, from create_index.
Merging is easier, when the upstream regression test files (create_index)
are as unmodified as possible, and any GPDB added tests are added to
separate files (bitmap_index). Also, the only planner vs. ORCA differences
in output were in the GPDB-added parts of this, so this allows removing
the ORCA alternative expected output file for the create_index test.