Reconnect 8.3 specific code in xlogdump
Now that PostgreSQL 8.3 has been merged we can enable the 8.3 specific
parts of xlogdump. There was a bug in the upstream code however that
included calls using struct members not available until 9.0 (coming in
commit 375369ac), disconnect this in the typical if/endif way. Since
xlogdump is abandonware due to the inclusion of pg_xlogdump in 9.3
theres not much use in submitting this upstream since we effectively
maintain this code.