Add a safe version of to_date(), which validates the input date
The regular to_date() function accepts invalid dates and uses
the input to calculate a valid date based on the difference in
the input. 2016-02-31 becomes 2016-03-02.

The new function verifies that the output and the input date
are the same, using the format string provides as second
parameter. A difference raises an error.

Solves Tracker #119368939