When looking up the distributed XID for a local XID, go straight to the SLRU.
We used to scan the list of LocalDistribXactData objects, in shared memory,
to find the distributed XID corresponding a local XID, during visibility
checks. That turns out to be unnecessary: at the point when we scanned the
list, the distributed log SLRU has already been updated, so we might as well
check that directly.

Thanks to @ashwinstar for pointing this out.