Simplify management of distributed transactions.
We used to have a separate array of LocalDistributedXactData instances, and
a reference in PGPROC to its associated LocalDistributedXact. That's
unnecessarily complicated: we can store the LocalDistributedXact information
directly in the PGPROC entry, and get rid fo the auxiliary array and the
bookkeeping needed to manage that array.

This doesn't affect the backend-private cache of committed Xids that also
lives in cdblocaldistribxact.c.

Now that the PGPROC->localDistributedXactData fields are never accessed
by other backends, don't protect it with ProcArrayLock anymore. This makes
the code simpler, and potentially improves performance too (ProcArrayLock
can be very heavily contended on a busy system).