Disable foreign key constraints from after row triggers.
Greenplum does not support foreign keys but users can still declare them and
store the constraints in the catalog. It was found that foreign key constraints
were being called in after row triggers at the end of ExecInsert, ExecDelete,
and ExecUpdate. ExecDelete's call to ExecARDeleteTriggers and ExecUpdate's call
to ExecARUpdateTriggers would result in a faulty tuple lookup and an ERROR. The
expected result is a noop trigger after the DML. Actually, we don't really
support triggers much either...