Invert enum range check to silence compiler warning
Since the actual datatype used for an enum is implementation specific
per each compiler in both C90 and C99 we can't rely on it automatically
being an unsigned integer even though it most likely will be in this
case. Inverting the check and testing for it's negation will achieve
the range check that was intended without compiler warnings issued for
comparison of unsigned enum (-Wtautological-compare in clang). Adding
a comment explaining this both in the code and the headerfile.

Note that the existing coding was actually incorrect as it was testing
for the tabletype ranges with logical AND and not logical OR which
would've been correct, the severity of this bug must be considered
quite low though.