Update error hint to include all supported formats
Include all the supported formats for the FORMAT option on external
tables in the errhint() call and perform minor copy editing to make
it conform to the error message guidelines. Also remove the list of
formats from the function comment since those lists are notorious
for getting outdated (case in point).