Make static compilation of ORCA work.
Reorder -lgpopt, -lnaucrates, -lgpdbcost, -lgpos, to make static
compilation of ORCA to work. When linking with dynamic libraries, the order
doesn't matter, but when with static libraries, the options need to be
given in dependency order.

This doesn't change the way those libraries are actually linked. Whether
we link with static or shared versions depend on which ones are found
at link time.