Backport src/test/isolation from PostgreSQL 9.1.
This brings in the directory exactly as it is in the upstream. It doesn't
compile yet, I will clean this up in a separate commit, to make it more
clear in the git history what the differences from upstream version are.

This doesn't include any of the upstream test specs, either. That's because
we don't support most of the features tested by those specs, yet. We'll get
the test scripts as we merge them from upstream, in due course, but for now,
we're going to use isolationtester only for our own tests.