Move GPDB-specific tests from 'alter_table' to 'alter_table_ao'.
Keeping upstream test files as close to upstream as possible helps with
diffing and merging with upstream. There are more GPDB-specific changes
there still, but this is a step in the right direction. The 'alter_table'
test runs for quite a long time, so by running it more in parallel, this
also shortens a installcheck-good run slightly.